refactor: 계산시 BigDecimal 타입 사용하도록 변경

+++ src/test/kotlin/calculator/ExpressionTest.kt
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
+import java.math.BigDecimal
 
 class ExpressionTest {
     @ParameterizedTest(name = "{0} = {1}")
@@ -15,7 +16,7 @@ class ExpressionTest {
         ],
         delimiter = '='
     )
-    fun `문자열로 이루어진 식을 받아 계산한다`(str: String, expected: Int) {
+    fun `문자열로 이루어진 식을 받아 계산한다`(str: String, expected: BigDecimal) {
         val expression = Expression.from(str)
         Assertions.assertThat(expression.getResult()).isEqualTo(expected)
     }
